backend/requirements.txt
asgiref==3.3.0
autopep8==1.5.4
Django==3.1.3
django-cors-headers==3.5.0
django-jazzmin==2.3.11
django-mysql==3.9.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
mysqlclient==2.0.1
pycodestyle==2.6.0
PyJWT==1.7.1
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
